+++ sdk/go/x/auto/errors_test.go
@@ -79,7 +79,7 @@ func TestConcurrentUpdateError(t *testing.T) {
 }
 
 func TestInlineConcurrentUpdateError(t *testing.T) {
-	t.Skip("disabled, see https://github.com/pulumi/pulumi/issues/5378")
+	t.Skip("disabled, see https://github.com/pulumi/pulumi/issues/5312")
 	ctx := context.Background()
 	pName := "inline_conflict_error"
 	sName := fmt.Sprintf("int_test%d", rangeIn(10000000, 99999999))